restricted all routes per user

+++ controllers/controllers.js
@@ -4,6 +4,7 @@ const { HttpError } = require("../helpers");
 const getAllContacts = async (req, res, next) => {
   try {
     const { _id: owner } = req.user;
+
     const result = await ContactModel.find({ owner }).populate("owner", "name email");
     res.json(result);
   } catch (err) {
@@ -13,8 +14,10 @@ const getAllContacts = async (req, res, next) => {
 
 const getById = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const { id } = req.params;
-    const result = await ContactModel.findById(id);
+
+    const result = await ContactModel.findById({ _id: id, owner });
     if (!result) {
       throw HttpError(404, "Not found");
     }
@@ -28,9 +31,10 @@ const addContact = async (req, res, next) => {
   try {
       
     const { _id: owner } = req.user;
-      const body = req.body;
+    const body = req.body;
       
-        const result = await ContactModel.create({...body, owner});
+    const result = await ContactModel.create({ ...body, owner });
+    
         res.status(201).json(result);
     } catch (err) {
         next(err);
@@ -39,10 +43,12 @@ const addContact = async (req, res, next) => {
 
 const updContactById = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const body = req.body;
-      const { id } = req.params;
+    const { id } = req.params;
       
-    const result = await ContactModel.findByIdAndUpdate(id, body, {new:true});
+    const result = await ContactModel.findByIdAndUpdate({ _id: id, owner }, body, { new: true });
+
     if (!result) {
       throw HttpError(404, "Not Found");
     }
@@ -54,9 +60,12 @@ const updContactById = async (req, res, next) => {
 
 const updFavorite = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const { id } = req.params;
     const body = req.body;
-    const result = await ContactModel.findByIdAndUpdate(id, body, {new:true});
+     
+
+    const result = await ContactModel.findByIdAndUpdate({ _id: id, owner }, body, { new: true });
 
     if (!result) {
       throw HttpError(400, "Missing favorite field");
@@ -68,17 +77,18 @@ const updFavorite = async (req, res, next) => {
 };
     
 const deleteContactById = async (req, res, next) => {
-    try {
-        const { id } = req.params;
+  try {
+    const { _id: owner } = req.user;
+    const { id } = req.params;
 
-        const result = await ContactModel.findByIdAndDelete(id);
-        if (!result) {
-            throw HttpError(404, "Not Found");
-        }
-        res.json({ message: "Delete was sucssesful" });
-    } catch (err) {
-        next(err);
+    const result = await ContactModel.findByIdAndDelete({ _id: id, owner });
+    if (!result) {
+      throw HttpError(404, "Not Found");
     }
+    res.json({ message: "Delete was sucssesful" });
+  } catch (err) {
+    next(err);
+  }
 };
 
 module.exports = {
